# docs/swagger.yaml
definitions:
  echo.HTTPError:
    properties:
      message: {}
    type: object
...
  handlers.BulkSendMessageRequest:
    properties:
      messages:
        description: List of messages to send
        items:
          $ref: '#/definitions/handlers.SendMessageRequest'
        type: array
    type: object
  handlers.BulkSendMessageResponse:
    properties:
      count:
        description: Number of messages accepted for processing
        example: 5
        type: integer
      message:
        description: Message indicating that bulk processing has started
        example: Bulk message processing started. Check your logs for more details.
        type: string
    type: object
  handlers.CreateAPIKeyResponse:
    properties:
      api_key:
        description: API key created
        example: ak_jkdfkjdfkdfjkdlklklklkllklklklklklklklklklkl
        type: string
      created_at:
        description: Timestamp of when the API key was created
        example: "2023-10-01T12:00:00Z"
        type: string
      description:
        description: Description of the API key
        example: This key is used for accessing the QDroid API.
        type: string
      expires_at:
        description: Expiration date for the API key
        example: "2024-12-31T23:59:59Z"
        type: string
      hashed_key:
        description: Hashed key of the created API key
        example: $2a$10$abcdefghijklmnopqrstuvwxyz
        type: string
      key_id:
        description: Key ID of the created API key
        example: ak_jkdfkjdfkdfjkd
        type: string
      message:
        description: Message indicating successful creation
        example: API key created successfully
        type: string
      name:
        description: Name of the API key
        example: My API Key
        type: string
    type: object
...
  handlers.GenericResponse:
    properties:
      message:
        description: Message indicating the result of the operation
        type: string
    type: object
...
  handlers.SendMessageRequest:
    properties:
      content:
        description: The message content to be sent
        example: Hello, World!
        type: string
      exchange_id:
        description: The exchange ID to send the message to
        example: ex_jkdfkjdfkdfjkd
        type: string
      phone_number:
        description: The phone number to send the message to
        example: "+2371234567890"
        type: string
      queue_id:
        description: The queue ID to use for sending the message
        example: exch_jkdfkjdfkdfjkd.237.62401
        type: string
    type: object
...
info:
  contact: {}
paths:
  /v1/auth/api-keys:
    post:
      description: Creates a new API key for the authenticated user.
      parameters:
      - default: Bearer <your_token_here>
        description: Bearer token for authentication. Replace <your_token_here> with
          a valid token.
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: API key created successfully
          schema:
            $ref: '#/definitions/handlers.CreateAPIKeyResponse'
        "400":
          description: Bad request, missing required fields
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "401":
          description: Unauthorized, invalid or expired session token
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "409":
          description: Duplicate API key name detected
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      security:
      - BearerAuth: []
      summary: Create API key
      tags:
      - auth
...
  /v1/messages/bulk-send:
    post:
      consumes:
      - application/json
      description: Sends multiple messages in bulk. Processing is done asynchronously.
      parameters:
      - default: Bearer <your_token_here>
        description: Bearer token for authentication. Replace <your_token_here> with
          a valid token.
        in: header
        name: Authorization
        required: true
        type: string
      - description: Bulk send message request payload
        in: body
        name: bulkSendMessageRequest
        required: true
        schema:
          $ref: '#/definitions/handlers.BulkSendMessageRequest'
      produces:
      - application/json
      responses:
        "202":
          description: Bulk message processing started
          schema:
            $ref: '#/definitions/handlers.BulkSendMessageResponse'
        "400":
          description: Bad request, missing required fields or empty messages array
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "401":
          description: Unauthorized, invalid or expired session token
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      security:
      - BearerAuth: []
      summary: Send multiple messages
      tags:
      - messages
  /v1/messages/send:
    post:
      consumes:
      - application/json
      description: Sends a single message to the specified exchange and phone number.
      parameters:
      - default: Bearer <your_token_here>
        description: Bearer token for authentication. Replace <your_token_here> with
          a valid token.
        in: header
        name: Authorization
        required: true
        type: string
      - description: Send message request payload
        in: body
        name: sendMessageRequest
        required: true
        schema:
          $ref: '#/definitions/handlers.SendMessageRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Message processed successfully
          schema:
            $ref: '#/definitions/handlers.GenericResponse'
        "400":
          description: Bad request, missing required fields or invalid phone number
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "401":
          description: Unauthorized, invalid or expired session token
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "404":
          description: Exchange or queue not found
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      security:
      - BearerAuth: []
      summary: Send a single message
      tags:
      - messages
...
swagger: "2.0"